requirements.txt
absl-py==0.7.1
astor==0.8.0
cycler==0.10.0
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
joblib==0.13.2
kiwisolver==1.1.0
Markdown==3.1.1
numpy==1.17.0
opencv-python==4.1.0.25
pandas==0.25.0
pathlib==1.0.1
Pillow==6.1.0
progressbar==2.5
protobuf==3.9.0
pydot==1.4.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.21.3
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
termcolor==1.1.0
torch==1.1.0
torchvision==0.3.0
Werkzeug==0.15.5
